feat: Update ability check UI for selected subjects
This commit updates the UI for the ability check section in InputInforPage.tsx. The colors of the subject boxes are now dynamically changed based on the values of globalData.ability. The "Toán" subject box background color is set to "#A3A3F2" if globalData.ability.math is greater than 0, and the text color is changed to black. Similarly, the "Hóa" subject box background color is set to "#A3A3F2" if globalData.ability.chemistry is greater than 0, and the text color is changed to black. This change improves the visual representation of the selected subjects in the UI.