feat: Update UI for ability check section in TopicDetail screen
This commit updates the UI for the ability check section in the TopicDetail screen of the DocsChosenTopic component. The colors of the subject boxes are now dynamically changed based on the values of globalData.ability. This change improves the visual representation of the selected subjects in the UI and enhances the user experience.